refactor(app): split SourceForm into smaller sibling files
Split the 2,929-line SourceForm.tsx into a Sources/SourceForm/ directory
of focused files. Pure file moves — no functionality or logic changes.

Largest file is now SourceForm.tsx at 832 lines (down from 2,929).
