+++ packages/app/src/components/Sources/SourceForm/index.ts
@@ -1,0 +1,1 @@
+export { TableSourceForm } from './SourceForm';
